attempting to split the 400 line file into ccps and hs
