--- PlatformIO/TeensyDevelopment/lib/ODriveTeensyCAN/README.md
@@ -1,1 +1,0 @@
-Originally from https://github.com/G-Levine/ODriveTeensyCAN
