--- PlatformIO/TeensyDevelopment/lib/mcp_can/keywords.txt
@@ -1,60 +1,0 @@
-#######################################
-# Syntax Coloring Map For debug_lvc
-#######################################
-
-#######################################
-# Datatypes (KEYWORD1)
-#######################################
-MCP_CAN	KEYWORD1
-mcp_can_dfs	KEYWORD1
-mcp_can	KEYWORD1
-
-#######################################
-# Methods and Functions (KEYWORD2)
-#######################################
-begin	KEYWORD2
-setMode	KEYWORD2
-init_Mask	KEYWORD2
-init_Filt	KEYWORD2
-sendMsgBuf	KEYWORD2
-readMsgBuf	KEYWORD2
-checkReceive	KEYWORD2
-checkError	KEYWORD2
-
-#######################################
-# Constants (LITERAL1)
-#######################################
-MCP_8MHZ	LITERAL1
-MCP_16MHZ	LITERAL1
-MCP_20MHZ	LITERAL1
-CAN_4K096BPS	LITERAL1
-CAN_5KBPS	LITERAL1
-CAN_10KBPS	LITERAL1
-CAN_20KBPS	LITERAL1
-CAN_31K25BPS	LITERAL1
-CAN_40KBPS	LITERAL1
-CAN_50KBPS	LITERAL1
-CAN_80KBPS	LITERAL1
-CAN_100KBPS	LITERAL1
-CAN_125KBPS	LITERAL1
-CAN_200KBPS	LITERAL1
-CAN_250KBPS	LITERAL1
-CAN_500KBPS	LITERAL1
-CAN_1000KBPS	LITERAL1
-MCP_ANY	LITERAL1
-MCP_STD	LITERAL1
-MCP_EXT	LITERAL1
-MCP_STDEXT	LITERAL1
-CAN_OK	LITERAL1
-CAN_FAILINIT	LITERAL1
-CAN_FAILTX	LITERAL1
-CAN_MSGAVAIL	LITERAL1
-CAN_NOMSG	LITERAL1
-CAN_CTRLERROR	LITERAL1
-CAN_GETTXBFTIMEOUT	LITERAL1
-CAN_SENDMSGTIMEOUT	LITERAL1
-CAN_FAIL	LITERAL1
-MCP_NORMAL	LITERAL1
-MCP_SLEEP	LITERAL1
-MCP_LOOPBACK	LITERAL1
-MCP_LISTENONLY	LITERAL1
