--- PlatformIO/TeensyDevelopment/lib/mcp_can/library.properties
@@ -1,9 +1,0 @@
-name=mcp_can
-version=1.5.1
-author=coryjfowler
-maintainer=coryjfowler
-sentence=MCP_CAN Library
-paragraph=Adds support for Microchip CAN 2.0B protocol controllers (MCP2515, MCP25625, and similar)
-category=Communication
-url=https://github.com/coryjfowler/MCP_CAN_lib
-architectures=*
