--- PlatformIO/TeensyDevelopment/lib/mcp_canbus/examples/send/send.ino
@@ -1,75 +1,0 @@
-/*  send a frame from can bus
-
-    CAN Baudrate,
-    
-    #define CAN_5KBPS           1
-    #define CAN_10KBPS          2
-    #define CAN_20KBPS          3
-    #define CAN_25KBPS          4 
-    #define CAN_31K25BPS        5
-    #define CAN_33KBPS          6
-    #define CAN_40KBPS          7
-    #define CAN_50KBPS          8
-    #define CAN_80KBPS          9
-    #define CAN_83K3BPS         10
-    #define CAN_95KBPS          11
-    #define CAN_100KBPS         12
-    #define CAN_125KBPS         13
-    #define CAN_200KBPS         14
-    #define CAN_250KBPS         15
-    #define CAN_500KBPS         16
-    #define CAN_666KBPS         17
-    #define CAN_1000KBPS        18
-*/
-   
-#include <mcp_canbus.h>
-#include <SPI.h>
-
-/* Please modify SPI_CS_PIN to adapt to your board.
-
-   CANBed V1        - 17
-   CANBed M0        - 3
-   CAN Bus Shield   - 9
-   CANBed 2040      - 9
-   CANBed Dual      - 9
-   OBD-2G Dev Kit   - 9
-   OBD-II GPS Kit   - 9
-   Hud Dev Kit      - 9
-
-   Seeed Studio CAN-Bus Breakout Board for XIAO and QT Py - D7
-*/
-
-#define SPI_CS_PIN  9 
-
-MCP_CAN CAN(SPI_CS_PIN);                                    // Set CS pin
-
-void setup()
-{
-    Serial.begin(115200);
-    while(!Serial);
-    
-    // below code need for OBD-II GPS Dev Kit Atemga32U4 version
-    // pinMode(A3, OUTPUT);
-    // digitalWrite(A3, HIGH);
-    
-    // below code need for OBD-II GPS Dev Kit RP2040 version
-    // pinMode(12, OUTPUT);
-    // digitalWrite(12, HIGH);
-    
-    while (CAN_OK != CAN.begin(CAN_500KBPS))    // init can bus : baudrate = 500k
-    {
-        Serial.println("CAN BUS FAIL!");
-        delay(100);
-    }
-    Serial.println("CAN BUS OK!");
-}
-
-unsigned char stmp[8] = {0, 1, 2, 3, 4, 5, 6, 7};
-void loop()
-{
-    CAN.sendMsgBuf(0x00, 0, 8, stmp);
-    delay(100);                       // send data per 100ms
-}
-
-// END FILE
-
