--- PlatformIO/TeensyDevelopment/lib/mcp_canbus/library.properties
@@ -1,9 +1,0 @@
-name=mcp_canbus
-version=1.0.0
-author=Longan Labs
-maintainer=Longan Labs
-sentence=A library for MCP2515
-paragraph=Arduino CAN Bus Library for MCP2515
-category=Communication
-url=https://github.com/Longan-Labs/Aruino_CAN_BUS_MCP2515
-architectures=*
