--- PlatformIO/TeensyDevelopment/lib/teensy_VESC_CANBUS/README.md
@@ -1,1 +1,0 @@
-Originally from https://github.com/skipper762/teensy_VESC_CANBUS
