--- PlatformIO/TeensyDevelopment/lib/teensy_VESC_CANBUS/datatypes.h
@@ -1,111 +1,0 @@
-/*
-        Copyright 2012-2016 Benjamin Vedder	benjamin@vedder.se
-
-        This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-    */
-
-#ifndef DATATYPES_H_
-#define DATATYPES_H_
-
-#include <stdbool.h>
-#include <stdint.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-// CAN commands
-typedef enum {
-  CAN_PACKET_SET_DUTY = 0,
-  CAN_PACKET_SET_CURRENT,
-  CAN_PACKET_SET_CURRENT_BRAKE,
-  CAN_PACKET_SET_RPM,
-  CAN_PACKET_SET_POS,
-  CAN_PACKET_FILL_RX_BUFFER,
-  CAN_PACKET_FILL_RX_BUFFER_LONG,
-  CAN_PACKET_PROCESS_RX_BUFFER,
-  CAN_PACKET_PROCESS_SHORT_BUFFER,
-  CAN_PACKET_STATUS,
-  CAN_PACKET_SET_CURRENT_REL,
-  CAN_PACKET_SET_CURRENT_BRAKE_REL,
-  CAN_PACKET_SET_CURRENT_HANDBRAKE,
-  CAN_PACKET_SET_CURRENT_HANDBRAKE_REL,
-  CAN_PACKET_STATUS_2,
-  CAN_PACKET_STATUS_3,
-  CAN_PACKET_STATUS_4,
-  CAN_PACKET_PING,
-  CAN_PACKET_PONG,
-  CAN_PACKET_DETECT_APPLY_ALL_FOC,
-  CAN_PACKET_DETECT_APPLY_ALL_FOC_RES,
-  CAN_PACKET_CONF_CURRENT_LIMITS,
-  CAN_PACKET_CONF_STORE_CURRENT_LIMITS,
-  CAN_PACKET_CONF_CURRENT_LIMITS_IN,
-  CAN_PACKET_CONF_STORE_CURRENT_LIMITS_IN,
-  CAN_PACKET_CONF_FOC_ERPMS,
-  CAN_PACKET_CONF_STORE_FOC_ERPMS,
-  CAN_PACKET_STATUS_5,
-  CAN_PACKET_POLL_TS5700N8501_STATUS,
-  CAN_PACKET_CONF_BATTERY_CUT,
-  CAN_PACKET_CONF_STORE_BATTERY_CUT,
-  CAN_PACKET_SHUTDOWN,
-  CAN_PACKET_IO_BOARD_ADC_1_TO_4,
-  CAN_PACKET_IO_BOARD_ADC_5_TO_8,
-  CAN_PACKET_IO_BOARD_ADC_9_TO_12,
-  CAN_PACKET_IO_BOARD_DIGITAL_IN,
-  CAN_PACKET_IO_BOARD_SET_OUTPUT_DIGITAL,
-  CAN_PACKET_IO_BOARD_SET_OUTPUT_PWM,
-  CAN_PACKET_BMS_V_TOT,
-  CAN_PACKET_BMS_I,
-  CAN_PACKET_BMS_AH_WH,
-  CAN_PACKET_BMS_V_CELL,
-  CAN_PACKET_BMS_BAL,
-  CAN_PACKET_BMS_TEMPS,
-  CAN_PACKET_BMS_HUM,
-  CAN_PACKET_BMS_SOC_SOH_TEMP_STAT,
-  CAN_PACKET_PSW_STAT,
-  CAN_PACKET_PSW_SWITCH,
-  CAN_PACKET_BMS_HW_DATA_1,
-  CAN_PACKET_BMS_HW_DATA_2,
-  CAN_PACKET_BMS_HW_DATA_3,
-  CAN_PACKET_BMS_HW_DATA_4,
-  CAN_PACKET_BMS_HW_DATA_5,
-  CAN_PACKET_BMS_AH_WH_CHG_TOTAL,
-  CAN_PACKET_BMS_AH_WH_DIS_TOTAL,
-  CAN_PACKET_UPDATE_PID_POS_OFFSET,
-  CAN_PACKET_POLL_ROTOR_POS,
-  CAN_PACKET_NOTIFY_BOOT,
-  CAN_PACKET_MAKE_ENUM_32_BITS = 0xFFFFFFFF,
-} CAN_PACKET_ID;
-
-// TODO implement other can status messages
-typedef struct {
-  float rpm;
-  float current;
-  // float amp_hours;
-  // float amp_hours_charged;
-  // float watt_hours;
-  // float watt_hours_charged;
-  // float temp_fet;
-  // float temp_motor;
-  // float current_in;
-  // float pid_pos_now;
-  float v_in;
-  int32_t tacho_value;
-} node_status;
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* DATATYPES_H_ */
